requirements.txt
mlx>=0.5.0
mlx-lm>=0.0.6
numpy
transformers
huggingface-hub
mpi4py
pyyaml
tqdm
